[#8] Requested changes: move FakeProducer to module kafka-test-utils.
