+++ bom/pom.xml
@@ -837,6 +837,12 @@
         <scope>test</scope>
       </dependency>
       <dependency>
+        <groupId>org.eclipse.hono</groupId>
+        <artifactId>kafka-test-utils</artifactId>
+        <version>${project.version}</version>
+        <scope>test</scope>
+      </dependency>
+      <dependency>
         <groupId>de.flapdoodle.embed</groupId>
         <artifactId>de.flapdoodle.embed.mongo</artifactId>
         <version>${flapdoodle.version}</version>
