+++ clients/adapter-kafka/pom.xml
@@ -43,6 +43,11 @@
 
         <!-- Testing -->
         <dependency>
+            <groupId>org.eclipse.hono</groupId>
+            <artifactId>kafka-test-utils</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
             <groupId>org.mockito</groupId>
             <artifactId>mockito-core</artifactId>
             <scope>test</scope>
