+++ clients/kafka-common/pom.xml
@@ -100,6 +100,11 @@
             <artifactId>assertj-core</artifactId>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>org.eclipse.hono</groupId>
+            <artifactId>kafka-test-utils</artifactId>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 
 </project>
