+++ test-utils/kafka-test-utils/src/test/java/org/eclipse/hono/kafka/test/FakeProducerTest.java
@@ -11,7 +11,7 @@
  * SPDX-License-Identifier: EPL-2.0
  */
 
-package org.eclipse.hono.kafka.client.test;
+package org.eclipse.hono.kafka.test;
 
 import static org.assertj.core.api.Assertions.assertThat;
