+++ test-utils/pom.xml
@@ -30,6 +30,7 @@
   <modules>
     <module>core-test-utils</module>
     <module>service-base-test-utils</module>
+      <module>kafka-test-utils</module>
   </modules>
 
   <properties>
